Clean up come bad code
